Move `imperativeModal`
