+++ apps/meteor/client/lib/2fa/process2faReturn.ts
@@ -1,5 +1,5 @@
 import { SHA256 } from '@rocket.chat/sha256';
-import { imperativeModal } from '@rocket.chat/ui-client';
+import { imperativeModal } from '@rocket.chat/ui-client/lib/imperativeModal';
 import { Meteor } from 'meteor/meteor';
 import { lazy } from 'react';
