--- packages/ui-client/.eslintignore
@@ -1,1 +1,0 @@
-!.storybook
