+++ packages/ui-client/src/helpers/index.ts
@@ -1,2 +1,1 @@
 export * from './getBaseURI';
-export * from './imperativeModal';
